Update for 2020 Update.
For this case, with new HW. We may not to add vibra sensor on board.
So add a config that if we use vibra sensor or not.
If there is no vibra sensor, we boot up directly to 10 days. once we press key, we reset the wake up remain to the default hours. 7 Days